add test case for call with macros
which matches a match found in the perl source.

This match in particular probably shouldn't be a hit, but how do
I differentiate this from other hits, possibly by both parameters
being macros, but there's not generic way to say "is this expression
a macro replacement" with specifying the macro name.

I might need to adapt the isExpandedFromMacro() check to see if we
find macro replacements that exactly match the expression bounds.

We can't just check if the expression is the result of any macro
expansion since it's always the result of the sv_setpvn() macro being
expanded and I don't want to hard code the WARNbits etc macro names
in the check.