make the implementation more general
I'm not entirely happy with the formatting mechanism used to build the
replacment call, in Perl I'd use something like:

  join(", ", map { getargtext($_) } @args);

but doing that requires a very late standard.

I could use accumulate(), but I think that would require turning every
StringRef into a std::string and more std::strings for each step of
the accumulate, which is expensive.

Of course, iostreams is expensive too.  I may re-work this soonish.